+++ muk_web_theme/models/__init__.py
@@ -17,6 +17,6 @@
 #
 ###################################################################################
 
-from . import res_users
 from . import res_company
+from . import res_users
 from . import res_config_settings