+++ muk_web_theme/models/res_company.py
@@ -23,6 +23,27 @@ class ResCompany(models.Model):
     
     _inherit = 'res.company'
     
+    #----------------------------------------------------------
+    # Database
+    #----------------------------------------------------------
+    
     background_image = fields.Binary(
         string="Apps Menu Background Image",
         attachment=True)
+    
+    default_sidebar_preference = fields.Selection(
+        selection=[
+            ('invisible', 'Invisible'),
+            ('small', 'Small'),
+            ('large', 'Large')
+        ], 
+        string="Sidebar Type",
+        default='small')
+    
+    default_chatter_preference = fields.Selection(
+        selection=[
+            ('normal', 'Normal'),
+            ('sided', 'Sided'),
+        ], 
+        string="Chatter Position", 
+        default='sided')